ai_backend/requirements.txt
albumentations==1.4.1
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.2.0
cycler==0.12.1
Cython==3.0.9
easydict==1.13
exceptiongroup==1.2.0
fastapi==0.110.0
flatbuffers==23.5.26
fonttools==4.49.0
grpcio==1.62.0
grpcio-tools==1.62.0
h11==0.14.0
humanfriendly==10.0
idna==3.6
imageio==2.34.0
importlib_resources==6.1.2
insightface==0.7.3
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.3
matplotlib==3.8.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
onnx==1.15.0
onnxruntime==1.17.1
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
packaging==23.2
pillow==10.2.0
prettytable==3.10.0
protobuf==4.25.3
pydantic==2.6.3
pydantic_core==2.16.3
pyparsing==3.1.1
pyreadline3==3.4.1
python-dateutil==2.9.0.post0
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.31.0
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
sympy==1.12
threadpoolctl==3.3.0
tifffile==2024.2.12
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.27.1
wcwidth==0.2.13
zipp==3.17.0
